Move btn module to it's own file
